Revert "Fixed a small Bug"

+++ seeraytracer-2/src/main/java/de/ur/iw/seeRaytracer/Scene.java
@@ -31,8 +31,8 @@ public class Scene {
     }
 
     public double getAverageTriangleSize(){
-      double sum = 0;
-      double count = 0;
+      int sum = 0;
+      int count = 0;
 
       for (var triangle : triangles) {
         sum += triangle.getTriangleSize();
